+++ Projects/vbr_core_examples/CB_014_xfit_premelt_extended.m
@@ -23,43 +23,46 @@
   VBR.in.elastic.anharmonic.dG_dP = 1.987; % GPa / GPa
 
 
-%% Define the Thermodynamic State %%
-
-  VBR.in.SV.T_K=700:50:1200;
+  %% Define the Thermodynamic State %%
+  VBR.in.SV.T_K=1200:5:1500;
   VBR.in.SV.T_K=VBR.in.SV.T_K+273;
   sz=size(VBR.in.SV.T_K); % temperature [K]
-  VBR.in.SV.P_GPa = 0.2 * ones(sz); % pressure [GPa]
+  VBR.in.SV.P_GPa = full_nd(2.5, sz); % pressure [GPa]
 
-  % this method requires the solidus.
-  % you should write your own function for the solidus that takes all the other
-  % state variables as input. This is just for illustration
-  dTdz=0.5 ; % solidus slope [C/km]
-  dTdP=dTdz / 3300 / 9.8 / 1000 * 1e9; % [C/GPa ]
-  VBR.in.SV.Tsolidus_K=1000+dTdP*VBR.in.SV.P_GPa;
+  Tn_cases = [.96, .98, 1.0, 1.1, 1.1, 1.1, 1.1];
+  phi_cases = [0., 0., 0., 0.1, 1, 2, 4]*0.01;
 
   % remaining state variables (ISV)
-  VBR.in.SV.dg_um=3.1 * ones(sz); % grain size [um]
-  VBR.in.SV.rho = 3300 * ones(sz); % density [kg m^-3]
-  VBR.in.SV.sig_MPa = 10 * ones(sz); % differential stress [MPa]
-  VBR.in.SV.phi = 0.01 * (VBR.in.SV.T_K >VBR.in.SV.Tsolidus_K); % melt fraction
-  VBR.in.SV.f = 1./logspace(-2,4,100); % frequency range
+  VBR.in.SV.dg_um=full_nd(.004*1e6, sz); % grain size [um]
+  VBR.in.SV.rho = full_nd(3300, sz); % density [kg m^-3]
+  VBR.in.SV.sig_MPa = full_nd(1, sz); % differential stress [MPa]
+  VBR.in.SV.f = 1; % 1 Hz
 
-%% CALL THE VBR CALCULATOR %%
-  [VBR] = VBR_spine(VBR) ;
 
-%% plot frequency dependence %%
-  figure;
-  subplot(1,3,1)
-  semilogx(1./VBR.in.SV.f,squeeze(VBR.out.anelastic.xfit_premelt.M(1,:,:)/1e9));
-  ylabel('M [GPa]'); xlabel('period [s]')
-  ylim([0,80])
+  f1=figure();
+  nTn = numel(Tn_cases);
+  VBR_results = struct();
 
-  subplot(1,3,2)
-  loglog(1./VBR.in.SV.f,squeeze(VBR.out.anelastic.xfit_premelt.Qinv(1,:,:)));
-  ylabel('Q^-1'); xlabel('period [s]')
-  ylim([1e-3,.1])
+  for iTn = 1:nTn
+        VBRi = VBR;
+        VBRi.in.SV.phi = full_nd(phi_cases(iTn), sz);
+        VBRi.in.SV.Tsolidus_K = VBR.in.SV.T_K / Tn_cases(iTn);
+        [VBRi] = VBR_spine(VBRi) ;
 
-  subplot(1,3,3)
-  semilogx(1./VBR.in.SV.f,1e-3*squeeze(VBR.out.anelastic.xfit_premelt.V(1,:,:)));
-  ylabel('V_s [km/s]'); xlabel('period [s]')
-  saveas(gcf,'./figures/CB_004_xfit_premelt.png')
+        results.Q = VBRi.out.anelastic.xfit_premelt.Q;
+        VBR_results(iTn) = results;
+
+        dname = [num2str(Tn_cases(iTn)), ', ', num2str(phi_cases(iTn))];
+        figure(f1)
+        if iTn > 1
+            hold all
+        end
+        plot(VBR.in.SV.T_K - 273, results.Q, 'displayname', dname, 'linewidth', 1.5)
+        legend('Location','eastoutside','title', '(Tn, phi)')
+        xlabel('Temperature [C]', 'fontsize', 12)
+        ylabel('Qs', 'fontsize', 12)
+        ylim([0, 200])
+
+
+  end
+