+++ release_notes.md
@@ -1,2 +1,5 @@
 # v1.1.3
 
+## New Features
+* add a `VBR_save` function for saving `VBR` structures 
+* add framework for handling temporary files in test suite