+++ release_notes.md
@@ -3,3 +3,7 @@
 ## New Features
 * add a `VBR_save` function for saving `VBR` structures 
 * add framework for handling temporary files in test suite
+* add convenience function, `full_nd`, to create filled arrays
+
+## Bug fix
+* fix for undefined behavior of pre-melt scaling at Tn == 1.0